Add empty AutoRepository

+++ backend/src/main/java/com/hardelele/ct/repositories/AutoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AutoRepository extends JpaRepository<AutoModel,Long> {
+public interface AutoRepository extends JpaRepository<AutoModel,Integer> {
 }
